:recycle: Move actor methods to archetype
