+++ src/modules/game/engine/archetypes/actor.ts
@@ -1,8 +1,53 @@
-import type { IGrid } from '../../typings'
+import type { Entity, World } from 'ape-ecs'
+import type { IGrid, IGridDirection } from '../../typings'
+import type Map3D from '../../common/map-3d'
 import { Cell3D } from '../../common/cell-3d'
-import Map3D from '../../common/map-3d'
+import Transform from '../components/transform'
+import Sprite from '../components/sprite'
+import Actor from '../components/actor'
+import MapCell from '../components/map-cell'
+import Hop from '../components/hop'
+import { randomColor, randomHop, randomString } from '../seed-dev'
 
 const HOP_Z_LEVELS = [0, 1, -1] as const
+
+export function createActor(world: World, grid: IGrid, map: Map3D): Entity {
+	return world.createEntity({
+		c: {
+			[Transform.key]: {
+				type: Transform.typeName,
+				...grid,
+			},
+			[Sprite.key]: {
+				type: Sprite.typeName,
+				texture: 'cube:0',
+			},
+			[Actor.key]: {
+				type: Actor.typeName,
+				userID: randomString([48, 57], 18),
+				username: randomString([32, 126], Math.floor(Math.random() * 12) + 3),
+				color: randomColor(),
+			},
+			[MapCell.key]: {
+				type: MapCell.typeName,
+				cell: new Cell3D({
+					map,
+					...grid,
+					properties: { solid: true, platform: true },
+				}),
+			},
+		},
+	})
+}
+
+export function hopActor(actor: Entity, direction?: IGridDirection) {
+	if (actor.has(Hop.typeName)) return // Already hopping
+	actor.addComponent({
+		type: Hop.typeName,
+		key: Hop.key,
+		...(direction || randomHop()),
+	})
+}
 
 export function reserveTarget(cell: Cell3D, target: IGrid): void {
 	cell.spread(target, { solid: true })
