+++ src/modules/game/engine/seed-dev.ts
@@ -1,42 +1,7 @@
-import { Entity, IEntityConfig, World } from 'ape-ecs'
+import type { Entity, World } from 'ape-ecs'
+import type { Direction, IGrid, IGridDirection } from '../typings'
 import Map3D from '../common/map-3d'
-import { Cell3D } from '../common/cell-3d'
-import Transform from './components/transform'
-import Sprite from './components/sprite'
-import Actor from './components/actor'
-import Hop from './components/hop'
-import MapCell from './components/map-cell'
-import Engine from '.'
-import { Direction, IGrid, IGridDirection } from '../typings'
-
-export function createActor(world: World, grid: IGrid, map: Map3D): Entity {
-	return world.createEntity({
-		c: {
-			[Transform.key]: {
-				type: Transform.typeName,
-				...grid,
-			},
-			[Sprite.key]: {
-				type: Sprite.typeName,
-				texture: 'cube:0',
-			},
-			[Actor.key]: {
-				type: Actor.typeName,
-				userID: randomString([48, 57], 18),
-				username: randomString([32, 126], Math.floor(Math.random() * 12) + 3),
-				color: randomColor(),
-			},
-			[MapCell.key]: {
-				type: MapCell.typeName,
-				cell: new Cell3D({
-					map,
-					...grid,
-					properties: { solid: true, platform: true },
-				}),
-			},
-		},
-	} as IEntityConfig)
-}
+import { createActor, hopActor } from './archetypes/actor'
 
 export function addActors(world: World, map: Map3D, count: number): Entity[] {
 	const entities: Entity[] = []
@@ -62,7 +27,7 @@ export function randomString(
 	return value
 }
 
-const COLORS: number[] = [
+const ACTOR_COLORS: number[] = [
 	0x119922,
 	0x99ff22,
 	0x00ff22,
@@ -72,7 +37,7 @@ const COLORS: number[] = [
 ]
 
 export function randomColor(): number {
-	return COLORS[Math.floor(Math.random() * COLORS.length)]
+	return ACTOR_COLORS[Math.floor(Math.random() * ACTOR_COLORS.length)]
 }
 
 export function createGridPool(
@@ -93,15 +58,6 @@ export function createGridPool(
 		}
 	}
 	return pool
-}
-
-export function hopActor(actor: Entity, direction?: IGridDirection) {
-	if (actor.has(Hop.typeName)) return // Already hopping
-	actor.addComponent({
-		type: Hop.typeName,
-		key: Hop.key,
-		...(direction || randomHop()),
-	})
 }
 
 export function randomHop(): IGridDirection {
@@ -140,8 +96,8 @@ export function hopTest(world: World, map: Map3D) {
 	}
 }
 
-export function seedGame(engine: Engine, map: Map3D) {
-	const actors = addActors(engine.world, map, 100)
+export function seedGame(world: World, map: Map3D) {
+	const actors = addActors(world, map, 100)
 	setInterval(() => {
 		hopActor(actors[Math.floor(Math.random() * actors.length)])
 	}, 250)
